+++ ResxFormatter/ISettings.cs
@@ -13,6 +13,7 @@
     {
         StringComparer Comparer { get; }
         bool RemoveDocumentationComment { get; }
+        bool RemoveXsdSchema { get; }
         bool SortEntries { get; }
     }