+++ ResxFormatterTests/Fake/FakeSettings.cs
@@ -8,6 +8,7 @@
     {
         public StringComparer Comparer { get; set; } = StringComparer.Ordinal;
         public bool RemoveDocumentationComment { get; set; }
+        public bool RemoveXsdSchema { get; set; }
         public bool SortEntries { get; set; }
     }
 }